source/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2  
pinecone-client==5.0.1
langchain-pinecone==0.2.0
langchain-experimental==0.3.0
